[Qt-ui] Upgrade to Qt5
(customized/tested on OSX only)

The reason for doing this is because on OS X Yosemite, dragging a ROM image onto the GUI will not load the image, due to the file drop URL not being a typical file name. There are 2 solutions to this problem. One is to integrate a call to an Objective C function that uses Foundation framework in order to convert back to a filename, but it's kind of complicated for someone like me, who is not well versed in OBJC++ and linking a call to a function between languages. The other solution, migrate to QT5, is much, much simpler.

There is a new variable in the bsnes makefile, QT_PREFIX. It can be modified for local customization of the QT5 location. The default location is customized for OSX Homebrew installation. More specifically, my homebrew installation also has QT4, and QT5 was put in said location to avoid conflicts. Your location may differ if you only have installed QT5.