[Make] Easy-customize QT4 location
The Problem: Prior to this commit, bsnes-classic defaulted on OS X to use /Library/Frameworks for QT, and made it unintuitive to easily change this location. Furthermore, it lacked the framework directory linker flag (-F) for potential non-standard locations (/Library/Frameworks being automatically included in the framework search path). moc and rcc locations had to be manually added if a user had them out of $PATH (say for QT5/4 coexist), along with manually changing the location of the frameworks..

The solution: Now, with this patch, the ability to uncomment a line in the high-level Makefile to specify an alternative QT location is available, and it will automatically deduce the moc/rcc binary locations and frameworks. If the line remains commented, the prior functionality is retained (/Library/Frameworks and no explicit moc/rcc location set).